requirements.txt
canvasapi==2.1.0
todoist-python==8.1.3
python-dotenv==0.15.0
pytz==2021.1